# lmdb: add finalizer for readonly Cursors; fix obsure SIGABRT