--- a/lmdb/txn.go
+++ b/lmdb/txn.go
@@ -323,7 +323,7 @@ func (txn *Txn) Del(dbi DBI, key, val []byte) error {
 func (txn *Txn) OpenCursor(dbi DBI) (*Cursor, error) {
 	cur, err := openCursor(txn, dbi)
 	if cur != nil && txn.readonly {
-		runtime.SetFinalizer(cur, (*Cursor).Close)
+		runtime.SetFinalizer(cur, (*Cursor).close)
 	}
 	return cur, err
 }