@bmatsuo 2015-11-02
These benchmark numbers are really weird. Somehow they seem consistent.