@bmatsuo 2015-11-02
Leaving aside the benchmark numbers for seemingly unrelated tests the (readonly) cursor benchmark is significantly slower, and that is probably correct. Cursor.Renew is the same speed. So that may be acceptable for readonly cursors. Though not benchmarked, update cursors shouldn't have a noticeable overhead.