@bmatsuo 2015-11-02
A finalizer on Cursor also means it is theoretically possible to cache Cursors in a sync.Pool. This should be benchmarked.