@bmatsuo 2015-11-18
These are the benchcmp results from the latest commit. Clearing the finalizer in Cursor.Close appears to have improved the results dramatically.